move CborReader.SkipValue() implementation to a dedicated source file
