+++ src/libraries/System.Security.Cryptography.Encoding/tests/Cbor/CborConformanceLevel.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+#nullable enable
 using System.Diagnostics;
 using System.Text;
