+++ src/libraries/System.Security.Cryptography.Encoding/tests/Cbor/CborTag.cs
@@ -2,6 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+#nullable enable
+
 namespace System.Formats.Cbor
 {
     /// <summary>
